fix: move cross-layer tests, add lifespan_started to schema key sets
